Add MIT license, update blog data types, enhance layout styles, and improve README documentation

+++ src/routes/blog/[blogId]/page.d.ts
@@ -6,7 +6,7 @@ export interface BlogData {
 		metadata: {
 			title: string;
 			description: string;
-			timestamp: bigint;
+			timestamp: int;
 		};
 	};
 }
